[IMP] crm{_iap_mine}, *: clean views from lead generation button
*: event_crm, mass_mailing_crm, website_crm

Remove the "generate leads" buttons from certain crm lead views
when coming from a stat button. Some create are also set to false.
The context key 'no_crm_lead_generation' is used to not use the
lead generation button. It is the cleanest way since the views are
spread over several modules.

It is both done in the kanban controller and in the list view
templates, where it is invisible if the key is in context.

Task-3107232