[FIX] mail: fix template preview layout
Enlarge selection fields of record and language in the template preview form so
that the labels are not truncated.

Technical note: the 2 selection fields were not aligned due to the presence in
one of them of a div with class "o_field_many2one_extra" that has a margin. The
only way we found to align them is to add a class on the field targeting that
inner div to set its margin to 0 (using important).

Task-3098782