[FIX] {sale_}crm: avoid merge duplicate leads that have multi-company error
Currently in CRM, opportunity 1 belongs to company A. It resembles
opportunity 2, which belongs to company B and already has a sale order
(also belongs to B). When CRON tries to assign a sales team, a sales person,
and merges opportunities, a match is found, and it attempts to merge them
both, but it returns the "Multi Company" error because the sale order
belongs to another company.

This error stops the process for all other duplicate opportunities. It
should only skip merging opportunities that have multi-company issues,
not all opportunities.

The above issue is solved by two things:

- "_allocate_leads_deduplicate" calls "_handle_salesmen_assignment" to set the
  sales team, and "_merge_opportunity" to merge duplicate opportunities. In
  cases of multi-company issues, we do not need to assign a team or merge
  duplicate opportunities. So, this commit added a new method "_merge_leads"
  that calls "_handle_salesmen_assignment" and "_merge_opportunity".

- In sale_crm, override the "_merge_opportunity" and it will return the
  opportunities that have a sale order and whose saleorder company is
  different than the first opportunity company, the first company's
  sale order, or the current company, and the returned opportunities are
  removed from "teams_data",  so do not process any kind of unlink or
  update operation on these opportunities.

task-2955805