[IMP] web_editor: floating bar and commandbar scss review
Prior to this commit, there was unnecessary scss for the floating toolbar
and the commandbar in the web editor.

This commit also fixes some design issues like alignement for the icons
in the list of the commandbar, box shadows, and the responsiveness
of the floating toolbar on mobile.

In a old commit (459d4e27a860f3f0308ee0984de6f789baa65424), the toolbar
dropdown had the display static added, but it causes problems to the
responsiveness.
This commit put the display back to dynamic on the dropdown and so, fixes
the dropdown not showing on mobile because of the overflow-x on auto.
The dynamic state also allows the dropdown to change direction (dropdown
to dropup) when there's not enought space for the content of the dropdown.

task-3087826