[FIX] website_sale: improve the working of the remove button
Steps to reproduce the bug:
- Open a product page in edit mode.
- Replace the image with one of your own.
- Click on the "Remove" button.
- Nothing happens and the "Remove" button has to be clicked one more
time in order to work properly.

The goal of this commit is to remove the need of the double clicking
operation in order to remove a media. To do so, the
"o_modified_image_to_save" class is removed from the image. Thanks to
that, no attachment is created for the modified image during the
"request_save" operation and the remove button now works properly.

task-3111601