[FIX] web: tooltip question mark trigger area
The question mark button in form view supposed to trigger the `tooltip`
box is difficult to reach since the triggering area is very small.

This commit increases the size of the `?` by adding the `btn` and
`btn-sm` class to widen the triggering area and respect the semantic.

However the padding was not increased because with the already increased
size, it would push neighboring labels/inputs.

The `d-inline` class is necessary to override the `inline-block`
behaviour of the `btn` class. Otherwise on small screen the `?` could
sometimes fall under their label.

task-3099734