[FIX] web: settings spacing issues
Some settings pages are using inconsistent xml structure messing up the
spacing between inputs and labels. The icons `o_doc_link` and the 
`fa-building-o` are now too close to their neighboring label.

This commit fixes the icons spacing by applying a margin-left to the 
`o_doc_link` class and the bootstrap class `ms-2` on the building icons.
I decided not to apply a general margin to the `fa-building-o` because
these icons are also used in different context.

The setting page rental settings had a weird checkboxes rendering
in Rent Online/Unavailability days. This commit applies the `form-check`
class on a parent div to render the checkboxes in column and apply BS5
guidelines [1]. This fix is made with the checkbox consistency
task in mind (task-3094083). 

This commit also fixes the messed up inputs and weird rendering on 
mobile view by applying `row` div parents and `col-lg-*` classes to some
labels that were missing this class. This fixes the rental page and the
messed up inputs in global settings / ReCAPTCHA. 


Task-3113372

[1]: https://getbootstrap.com/docs/5.0/forms/checks-radios/#checks