[FIX] web: Replace missing fields widgets by a "ghost" widget
Steps to reproduce
==================

 1. Force modify a view with an unexisting field
    For example `res.users.tree` with:
    `<field name="unexisting_field"/>`
    ( note that odoo won't allow to do so,
      you will have to do it in raw PSQL * )
 2. Open Settings > User & Companies > Users
=> JS traceback error:

```js
UncaughtPromiseError > TypeError
Uncaught Promise > Cannot read properties of undefined (reading 'string')

TypeError: Cannot read properties of undefined (reading 'string')
    at http://localhost:8069/web/assets/debug/web.assets_backend.js:67523:84 (/web/static/src/legacy/legacy_load_views.js:67)
    ...
```
* : In practice, it happens to Odoo customers after
incorrect installations/removal of apps

Solution
========

Replace missing fields by a custom widget + log a message with more
details

This works by deserializing the xml, setting the widget of missing fields to the ghost widget and reserializing the arch.

There are some special cases we need to handle when checking if a field
is missing:

With the following arch:

```xml
<form>
   <field name="parent">
      <tree>
         <field name="child"/>
      </tree>
   </field>
</form>
```

When checking if the child field is missing, we need to check if it is
present in the model of the parent field.

The `groupby` also changes the current model.

TODO: add tests

opw-3081124