[FIX] web_editor: fix shape flipping and colors
This commit fixes 2 bugs with predefined shapes in theme snippets.

The first fix fixes an issue that causes thin spacing to appear between
two snippets containing shapes. This bug had already been fixed in this
commit [1] but the theme snippets have not been adapted to the changes.

With this commit, the shapes of theme snippets are automatically adapted
when they are dropped on the page.

Steps to reproduce the bug:

- Choose the "Clean Theme" for a website.
- Drag and drop a "Call to action" snippet on the page.
- Drag and drop a "Text" snippet with a dark background before it.
- Resize the window to change the window width.
- At some points, the gap will appear.

The second fix fixes an issue that appeared because of this same commit
[1].

Steps to reproduce the bug:

- Choose the "Clean Theme" for a website.
- Drag and drop a "Call to action" snippet on the page.
- Click on one of the two buttons to flip the shape of this snippet.
- Bug => The shape disappears.

This was due to shape colors set by theme variables in some theme
snippets (e.g. 'o-color-1'). When the background image url was recreated
to apply the "flip", the color was not converted to hexadecimal.

[1]: https://github.com/odoo/odoo/commit/42b3ad10e0b32b7fc72f801e2c67d6baf938c566

task-2824607
opw-3069213
opw-3057533