[IMP] onboarding: enable onboarding step sharing
Purpose: Allow to use the same step (and progress) for multiple onboardings.

This is done by mutating the onboarding - onboarding step relation to m2m.
This also required to transform the progress models relationship to m2m:

* Progress step records are now linked to several progress records if the step
they track is used in several onboardings.
* There will still be multiple progress step records for per-company tracking,
if applicable.

Tests are added and updated.

Task-3025136