[IMP] core: make some cookies expire on logout
Some cookies were left around even when the user logged out.

For `cids` the next user would not log into the default company but
instead connect into the company of the last user.

For `color_scheme` the next user would not use the default light mode in
case the last user was in dark mode.

task-3077421