[FIX] event_sale: convert displayed price on sale order
In the forward port of [this fix on pricelists](https://github.com/odoo/odoo/commit/7ef09a18a304d8ff5266a3af8f0ee12a56a1ec2f)
to [saas-15.3](https://github.com/odoo/odoo/commit/2a023ca1faa3667026ce1fabc5b35039ae7565fe)
there was a fix on the displayed price on the sale order which was not
converted. When adding an event in a sale order, the price of this event
was not converted to the selected pricelist. This commit solves this
issue for saas-15.2

opw-2916904