[IMP] account: bank statement smartbutton
When a user imports a bank statement, reconciles some moves,.. later on,
he does  not  have the ability to easely  go  back to the bank statement
and see the others items / continue his work.

This commit enable the user to move from the journal entry to the bank
statement (if any).

task-id: 3116184
[enterprise](https://github.com/odoo/enterprise/pull/35580)