[FIX] mail: prevent sanitation of URL parameters
Before this change URLs would get their url-safe encoding replaced
with regular characters when cleaning up
html content.

This caused signed URLs that use some of those characters to become
invalid if the target did not accept equivalent characters.

Concrete example:

/markers=size%3Alittle%7CAddress
is not the same as
/markers=size%3Alittle|Address
for google maps signed static urls

## Fix

We look for URLs in the 'dirty' document
Then match urls again in the cleaned document
As all urls should be preserved we can replace them 1 by 1.

task - 3079113