[IMP] stock*: split transfers menu in three
In the inventory menu, there is a "Transfers" menu item who displays all
the transfers.
This commit replaces it by three other menu items: "Receipts",
"Deliveries" and "Internal Transfers".
Each of them will display only the transfers with the matching picking's
type (depending of the picking type's code).
When a picking is created from there, its picking type will be set by
default according to it and the field will be filtered by its code.

task-3076044