[FIX] website_sale: more specific warning message on time based price products compatibility
Steps to reproduce:

- Create 2 products with recurring invoice, one with a monthly price and
one with a yearly price.
- Go to the shop in the frontend, and add one of them to the cart.
- Now go to the product page of the other product.

Issue:

We are going to see always the same warning message telling us that
'This product has no valid combination.' which is not very specific
about what is the reason we cannot add the product to the cart.

Solution:

Added a specific message for the products with recurring invoice that
can't be added to the cart.

opw-3084520