[IMP] project: improve ux
Before this commit:
- When a project is created on SO confirmation,
the new project (assuming there is no project template
set on the product) is added with one stage with name=New
- Project sharing > task description: checkboxes are
displayed as clickable.

After this commit:
- When a project is created on SO confirmation,
set the following stages for the new project
(assuming there is no project template set on the product):
to do - in progress - done (folded) - canceled (folded).
- Project sharing > task description: checkboxes are
no more displayed as clickable.
- the content of the task rating acknowledgment
page was centered vertically.

task-2929871