[FIX] website: fix link_tools runbot random fail
The tour 'link_tools' fails randomly when runned multiple times. this is
caused by the edition of links that were already in the page (using link
tools).

Steps:

- On edit mode > From the link tools, change link style to secondary,
- Click on 'save' > the button keeps the old style.

The goal of this commit, is to add an optional delay before 'Save' so
the process of '_setObserver' is completed correctly since (when doing
automated testing) the page "save" can be executed debfore marking the
DOM as dirty.

runbot-5753