[FIX] account: restrict display of 'reconciled entries' button on moves
When resetting a payment/move to draft, the related entries are unreconciled. The 'reconciled entries' button raises an error since there isn't any.

It doesn't make sense to have this button in this situation, so we hide it when the payment/move is in draft. However, resetting to draft then posting won't re-reconcile the entries, and the button will still raise an error. To avoid that, we also restrict it to moves that actually have reconciled entries.

task id=3120327