[FIX] hr_org_chart: badly defined computed fields
The computed fields 'subordinate_ids' and 'is_subordinate' suffered from
complex definition and incomplete dependencies.  This showed up in the
test, which had to explicitly invoke a compute method to test a field's
value.

First, the fields 'subordinate_ids' must be of type many2many instead of
one2many, simply because an employee can be the subordinate of more than
one employee.

Second, 'subordinate_ids' must be declared recursive, just like its
co-computed field 'child_all_count'.  And that makes sense: changing the
parent of an employee also impacts the subordinates of all its indirect
parents (ancestors).  I took the liberty to simplify the code to compute
'subordinate_ids', which is now a recursive one-liner.

Third, the dependencies of the field 'is_subordinate' were not correct
either.  If you change the parent of one of your subordinates, it will
no longer be your subordinate, and the same applies for its own
subordinates.  The field should be considered recursive for its
dependencies to be correct, and this makes the hack in the test no
longer necessary.