[REF] account,base_hash,l10n_fr{,_pos_cert}: refactor the hash system into a new mixin
Currently, the same chained hash system is duplicated in multiple modules (`account` and `l10n_fr_pos_cert`, but will also be used later in `l10n_pt`). It is also very rigid and cannot be easily reused (for instance to modify the fields used in the hash).

To make it more reusable, we now introduce a new module `base_hash` which holds a `hash.mixin`. This mixin that can be inherit by other models that will have to implement a few methods (e.g. the fields used for the hash, or when a model has to be hashed) so that they can use this chained hash mechanism more easily.

Moreover, this also allows us to generify the hash integrity report to avoid code duplication.

Task 2794952